Move formatLabel in labels.ts
